Refactor addbook method
